Refactor newAtomicOps to clean up ConstructionParse
